+++ cognite/experimental/_api/pnid_parsing.py
@@ -42,17 +42,12 @@ class PNIDParsingAPI(ContextAPI):
         Returns:
             PNIDDetectResults: Resulting queued job. Note that .results property of this job will block waiting for results."""
 
-        if not (
-            all([isinstance(entity, str) for entity in entities])
-            or all([isinstance(entity, dict) for entity in entities])
-            or all([isinstance(entity, CogniteResource) for entity in entities])
-        ):
-            raise ValueError("all the elements in entities must have same type (either str or dict)")
-
         if file_id is None and file_external_id is None:
             raise ValueError("File id and file external id cannot both be none")
 
-        entities, entities_field_to_objects = self._detect_before_hook(entities, search_field)
+        entities = [
+            entity.dump(camel_case=True) if isinstance(entity, CogniteResource) else entity for entity in entities
+        ]
 
         job = self._run_job(
             job_path="/detect",
@@ -61,27 +56,12 @@ class PNIDParsingAPI(ContextAPI):
             file_external_id=file_external_id,
             entities=entities,
             partial_match=partial_match,
+            search_field=search_field,
             name_mapping=name_mapping,
             min_tokens=min_tokens,
             job_cls=PNIDDetectResults,
         )
-        job._set_entities_field_to_objects(entities_field_to_objects)
         return job
-
-    @staticmethod
-    def _detect_before_hook(entities, search_field):
-        """To decide whether to use search_field or not and make sure the entities are of type List[str]"""
-        if entities and isinstance(entities[0], CogniteResource):
-            entities = [entity.dump(camel_case=True) for entity in entities]
-
-        if entities and isinstance(entities[0], dict):
-            entities_field_to_objects = defaultdict(list)
-            converted_entities = [entity.get(search_field) for entity in entities]
-            for entity_field, entity_obj in zip(converted_entities, entities):
-                entities_field_to_objects[entity_field].append(entity_obj)
-            return converted_entities, entities_field_to_objects
-        else:
-            return entities, None
 
     def extract_pattern(
         self, patterns: List[str], file_id: int = None, file_external_id: str = None